Add support for only including body in sigs if it has the correct content-type
Add the ability to specify a content_type when generating signatures or
checking the validity of a request.

The default is set to u'application/x-www-form-urlencoded' which will
attempt to parse any body content passed in.

If it is set to anything else the body won't be included in the
generation of the signature.

This brings it into line with Section 3.4.1.3.1. (Parameter Sources) in
rfc5849.