Oauth1 signing: fix error message on invalid Content-Type header
Content-Type should be application/x-www-form-urlencoded when the
request to sign has a body. If it's not the case a ValueError was raised
saying Content-Type was missing, while it actually could just be wrongly
set (e.g. to application/json). Fix the error message to reflect the
real problem: expected & given Content-Type.